feat: implement navbar component with shared configuration and styling for web and native environments
