+++ libs/navbar/src/index.ts
@@ -1,2 +1,11 @@
-export * from './lib/navbar';
-export * from './lib/shared-config';
+// Export all types
+export * from './types';
+
+// Export configuration
+export * from './config';
+
+// Export utilities
+export * from './utils';
+
+// Export web component
+export * from './components/web/navbar';
